+++ backend/authMiddleware.js
@@ -1,0 +1,14 @@
+const jwt = require("jsonwebtoken");
+
+const verificarToken = (req, res, next) => {
+    const token = req.headers["authorization"];
+    if (!token) return res.status(403).json({ error: "Acceso denegado" });
+
+    jwt.verify(token.split(" ")[1], process.env.SECRET_KEY, (err, decoded) => {
+        if (err) return res.status(401).json({ error: "Token inválido" });
+        req.user = decoded;
+        next();
+    });
+};
+
+module.exports = verificarToken;
