+++ backend/db.js
@@ -1,0 +1,12 @@
+const { Pool } = require("pg");
+require("dotenv").config();
+
+const pool = new Pool({
+    user: process.env.DB_USER || "postgres",
+    host: process.env.DB_HOST || "localhost",
+    database: process.env.DB_NAME || "tienda_tecnologica",
+    password: process.env.DB_PASSWORD || "12345",
+    port: process.env.DB_PORT || 5432,
+});
+
+module.exports = pool;
