+++ frontend/src/main.jsx
@@ -3,16 +3,16 @@ import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import "./index.css";
 import { BrowserRouter } from "react-router-dom";
-import { AuthProvider } from "./context/AuthContext";  // ✅ Debe ir primero
+import { AuthProvider } from "./context/AuthContext";  
 import { CartProvider } from "./context/CartContext";  
 import { HistoryProvider } from "./context/HistoryContext";  
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <AuthProvider>  {/* ✅ Primero el AuthProvider */}
-        <CartProvider>  {/* ✅ Luego el CartProvider */}
-          <HistoryProvider>  {/* ✅ Luego el HistoryProvider */}
+    <BrowserRouter basename="/">
+      <AuthProvider>  
+        <CartProvider>  
+          <HistoryProvider>  
             <App />
           </HistoryProvider>
         </CartProvider>
