+++ frontend/vite.config.js
@@ -3,15 +3,14 @@ import react from "@vitejs/plugin-react";
 
 export default defineConfig({
   plugins: [react()],
-  esbuild: {
-    loader: "jsx",
-  },
   server: {
     watch: {
       usePolling: true,
     },
+    historyApiFallback: true, // 🔥 Evita errores 404 en recargas de páginas dentro de React
   },
   build: {
+    outDir: "dist",
     target: "esnext",
   },
 });
